+++ include/roadmap.hpp
@@ -10,9 +10,11 @@ enum Profession {
     ALCHEMIST     = 3,
     FISHER        = 4,
     LOWERING_PODS = 5,
-    TRAIN         = 6,
-    GHOST         = 7,
-    MISSIONS      = 8
+    TRAIN         = 10,
+    TRAIN_OPENCV  = 11,
+    MISSIONS      = 20,
+    GHOST         = 90,
+    TEST          = 99
 };
 
 class Roadmap {
@@ -42,6 +44,19 @@ class Roadmap {
         int Start();
 
 
+        /**
+        * @brief Execute a single roadmap file (.csv) by visiting every map,
+        *        clicking the recorded identities, and change map.
+        *
+        * @param name File name of the csv telemetry for the roadmap.
+        * @return int Status code.
+        * Public because debug.
+        */
+        int ExecuteRoadMap(std::string name);
+
+
+        int ExecuteTrainingRoadMap(std::string name);
+
     private:
 
         Profession _profession;
@@ -64,17 +79,11 @@ class Roadmap {
             GO_TO_INITIAL_MAP           = 6,
             EXECUTE_ROADMAP             = 7,
             TRAIN_MODE                  = 8,
-            END_ROADMAP_OK              = 9
+            TRAIN_OPENCV_MODE           = 9,
+            END_ROADMAP_OK              = 10
         };
 
-        /**
-        * @brief Execute a single roadmap file (.csv) by visiting every map,
-        *        clicking the recorded identities, and change map.
-        *
-        * @param name File name of the csv telemetry for the roadmap.
-        * @return int Status code.
-        */
-        int ExecuteRoadMap(std::string name);
+
 
         /**
         * @brief Click all resource identities in a map (except last which is the