+++ include/train.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+#include <vector>
+#include <string>
+
 /**
  * @file train.hpp
  * @brief Helpers used by the training logic to move around maps and search cells where the enemies are.
@@ -29,8 +32,12 @@ class Train {
          */
         int IterateCells();
 
+
+        int FindMob(std::string mobName, std::vector<int> mobNumber);
+
     private:
 
         int CheckMobSize(int &x, int &y);
+        int ReadyToFight();
 
 };