+++ src/menusInterface.cpp
@@ -54,6 +54,7 @@ void show::WoodMenu()
     std::cout << "1  = lv.1  [ASTRUB]        (Astrub)" << std::endl;
     std::cout << "2  = lv.20 [AMAKNA]         (Linde)" << std::endl;
     std::cout << "4  = lv.70 [BOMBU]        (Llanura)" << std::endl;
+    std::cout << "5  = lv.90 [OLVIO]      (Caravanas)" << std::endl;
     std::cout << "8  = [OTOMAI | Coast]     (Costero)" << std::endl;
     std::cout << "9  = [OTOMAI | ALL]       (Costero)" << std::endl;
     std::cout << "10 = [KOALAK]           (Ganaderos)" << std::endl;
@@ -123,7 +124,7 @@ void show::FindEnemiesMenu()
 {
     system("TITLE Find enemies menu");
     std::cout << "===================================" << std::endl;
-    std::cout << "1  = Jabatos               (Astrub)" << std::endl;
+    std::cout << "1  = Bosque                (Astrub)" << std::endl;
     std::cout << "2  = Abraknidos             (Linde)" << std::endl;
     std::cout << "5  = Muluscos               (Golfo)" << std::endl;
     std::cout << "8  = Sidimonte          (Caravanas)" << std::endl;