+++ src/system/file.cpp
@@ -121,6 +121,7 @@ std::vector<std::vector<std::pair<int, int> > > File::ReadFileAndBuildMap(const 
         }
         ifs.close();
     } else {
+        File::LogFile("File doesnt exist... ", true);
         std::cout << "File doesnt exist: " << filePath << " / " << error.message() << std::endl;
         std::cout << "Creating new file..." << std::endl;
         std::filesystem::create_directories(path);
@@ -128,6 +129,17 @@ std::vector<std::vector<std::pair<int, int> > > File::ReadFileAndBuildMap(const 
 
     if(ret.size() == 0) {
         File::LogFile("Careful, empty file ? ", true);
+    }
+
+    return ret;
+}
+
+bool File::ExistFile(const std::string &filePath)
+{
+    bool ret = false;
+
+    if(std::filesystem::exists(filePath)) {
+        ret = true;
     }
 
     return ret;
@@ -150,6 +162,7 @@ std::string File::ReadFile(const std::string &filePath)
         }
         ifs.close();
     } else {
+        File::LogFile("File doesnt exist... ", true);
         std::cout << "File doesnt exist: " << filePath << " / " << error.message() << std::endl;
         std::cout << "Creating new file..." << std::endl;
         std::ofstream ofs(path);
@@ -176,6 +189,7 @@ void File::DeleteLocalFile(const std::string &filePath)
 
     std::error_code error;
     if(!std::filesystem::remove(path, error)) {
+        File::LogFile("Could not remove the file... ", true);
         std::cout << "! Could not remove the file: " << filePath << ". Error: " << error.message() << std::endl;
     }
 }